fix(controller/, api/profiles): renamed from services to controller
