+++ app/server/routes/api/profiles.ts
@@ -1,6 +1,6 @@
 import express, { Request, Response, NextFunction } from 'express';
 import createError from 'http-errors';
-import { registerUser } from '../../services/registration';
+import { registerUser } from '../../controller/registration';
 import { RegisterUserResponse } from '../../types/profiles.types';
 import hashPassword from '../../middleware/profiles/hash-password';
 import emailValidator from '../../middleware/profiles/email-validator';
